Don't run apt-get update from recipes
Running apt-get update inside recipes can cause problems if multiple
Chef processes are running in parallel, which is the case for cloudinit.d
launch plans with replicas. The apt database will be locked and the
apt-get process will exit with status 100.

Instead, use the apt recipe at the beginning of the basenode run_list.